WIP fix concurrency issues and TLS support
work in progress: works reliably (just as before) also when
TLS is enabled for the pubsubclient mqtt connection. This is the microphone audio
transmission to Rhasspy via hermes mqtt
Two more semaphore usage to synchronize
access to objects shared across tasks introduced.

As a positive side effect even for none-TLS based
 operation WIFI connects in my tests directly, no
reboot or wait happening.

Additional changes:
- use of hotword vs idle brightness: Except when
using COLOR_IDLE, brightness is based on hotword
brightness.
- tracking of ASR/HOTWORD toggleOn/toggleOff
state, used to decide when to go to Idle state (if
not saying anything and no ASR is running).

How to test TLS:
- Make sure your MQTT server speaks TLS at port 8883. Connection to port 1883 (or whatever was
set in config) still required for some MQTT
messages

- set MQTT_TLS to 1 (e.g. via build_flags)

- Build and flash. You should now get connected as normal.
Only difference is, that all audio FROM the ESP is
encrypted.